==> notebooks/evaluation/12_Features_Ablation_Study.ipynb <==
import os
from pathlib import Path
import pandas as pd
from src import utils
import lightgbm
import math
import re
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.neighbors import KNeighborsRegressor
from sklearn.neural_network import MLPRegressor
from sklearn.linear_model import Lasso, Ridge, LinearRegression, ElasticNet
from sklearn.svm import SVR
from tqdm import tqdm

# Project path
project_dir = str(Path().resolve().parents[1])
# Load enviromental variables
env_var = utils.load_env_variables(project_dir)

MAP_MODELS = {
    "LGBM": lightgbm.LGBMRegressor,
    "DT": DecisionTreeRegressor,
    "SVM": SVR,
    "KNN": KNeighborsRegressor,
    "MLP": MLPRegressor,
    "RF": RandomForestRegressor,
    "Lasso": Lasso,
    "OLS": LinearRegression,
    "Ridge": Ridge,
    "ElasticNet": ElasticNet,
}

def get_model(ml_method, n_features, params):
    """Get the models by name"""
    if ml_method == "KNN":
        return MAP_MODELS[ml_method](
            n_neighbors=3
        )
    if ml_method == "MLP":
        return MAP_MODELS[ml_method](
            (math.floor(1),),
            max_iter=10000
        )
    if ml_method == "RF":
        return MAP_MODELS[ml_method](n_estimators=200)
    if ml_method == "DT":
        return MAP_MODELS[ml_method]()
    if ml_method == "Lasso":
        return MAP_MODELS[ml_method]()
    if ml_method == "OLS":
        return MAP_MODELS[ml_method]()
    if ml_method == "Ridge":
        return MAP_MODELS[ml_method]()
    if ml_method == "ElasticNet":
        return MAP_MODELS[ml_method]()
    if ml_method == "SVM":
        return MAP_MODELS[ml_method]()
    return MAP_MODELS[ml_method](*params)


dataset = "Brazil_Election_2018"
validation = "CrossValidation"
ml_methods = ["KNN"]
fs_method = "CFS"
index_col = "INDEX"

data_path = os.path.join(env_var["root_path"], dataset, "Original", "data.csv")
fold_path = os.path.join(env_var["root_path"], dataset, "Original", "folds", validation)
features_path = os.path.join(env_var["root_path"], dataset, "Original", "results", validation, "features_selected", fs_method)

data = pd.read_csv(data_path, index_col=index_col)
data[["[GEO]_LATITUDE", "[GEO]_LONGITUDE"]]=(data[["[GEO]_LATITUDE", "[GEO]_LONGITUDE"]]-data[["[GEO]_LATITUDE", "[GEO]_LONGITUDE"]].min())/(data[["[GEO]_LATITUDE", "[GEO]_LONGITUDE"]].max()-data[["[GEO]_LATITUDE", "[GEO]_LONGITUDE"]].min())
data

from genericpath import isfile
from ntpath import join

results_method = {}
folds = [f for f in os.listdir(fold_path) if not os.path.isfile(os.path.join(fold_path, f))]
for ml_method in tqdm(ml_methods):
    features_abblation = []
    mse_abblation = []
    features_all = []
    for ablation_perc in tqdm([0]):
        mse_by_fold = []
        len_features = []
        for fold in folds:
            splited_data = utils.load_json(os.path.join(fold_path,fold,"split_data.json"))
            features = utils.load_json(os.path.join(features_path,f"{fold}.json"))
            if ablation_perc > -1:
                init_pointer = int(ablation_perc*len(features["selected_features"]))
                features["selected_features"] = features["selected_features"][: len(features["selected_features"]) - init_pointer]
                features["selected_features"] = features["selected_features"] + ["[GEO]_LATITUDE", "[GEO]_LONGITUDE"]
            X_train = data.loc[splited_data["train"],features["selected_features"]].copy()
            X_test = data.loc[splited_data["test"],features["selected_features"]].copy()
            y_train = data.loc[splited_data["train"],"TARGET"].copy()
            y_test = data.loc[splited_data["test"],"TARGET"].copy()

            clean_cols = [re.sub(r"\W+", "", col) for col in X_train.columns]
            X_train.columns = clean_cols
            X_test.columns = clean_cols

            model = get_model(ml_method, len(X_train.columns), {})
            model.fit(X_train,y_train)
            pred = model.predict(X_test)
            mse_by_fold.append(((y_test - pred)**2).mean())
            len_features.append(len(X_test.columns))
            features_all = features_all + features["selected_features"]
        features_abblation.append(sum(len_features)/len(len_features))
        mse_abblation.append(sum(mse_by_fold)/len(mse_by_fold))
    results_method[ml_method] = mse_abblation

features_all_folds = list(set(features_all))
features_all_folds.append("TARGET")
data_state = data[data["INDEX_FOLDS"]==23]
corr = data_state[features_all_folds].corr()
#corr.style.background_gradient(cmap='coolwarm')

pd.set_option('display.max_rows', 500)
corr.loc["TARGET"]

results = pd.DataFrame(results_method)
#results.to_csv("results.csv", index=False)
#results

(results.loc[1] - results.loc[0]).plot()
